# server/docs/swagger.yaml
basePath: /api
definitions:
...
  models.DailyMilk:
    properties:
      date:
        allOf:
        - $ref: '#/definitions/models.DateOnly'
        description: Дата дойки
      fat:
        description: Суммарный жир `example:"12"`
        type: integer
      fatEvening:
        description: Жир вечером `example:"12"`
        type: integer
      fatMorning:
        description: Жир утром`example:"12"`
        type: integer
      fatNoon:
        description: Жир днем `example:"12"`
        type: integer
      id:
        type: integer
      lactationId:
        description: ID лактации во время котороый была дойка `example:"1"`
        type: integer
      milk:
        description: Суммарный надой `example:"12"`
        type: integer
      milkEvening:
        description: Надой вечером `example:"12"`
        type: integer
      milkMorning:
        description: Надой утром `example:"12"`
        type: integer
      milkNoon:
        description: Надой днем `example:"12"`
        type: integer
      protein:
        description: Суммарный белок `example:"12"`
        type: integer
      proteinEvening:
        description: Белок вечером `example:"12"`
        type: integer
      proteinMorning:
        description: Белок утром `example:"12"`
        type: integer
      proteinNoon:
        description: Белок днем `example:"12"`
        type: integer
    type: object
  models.DateOnly:
    properties:
      time.Time:
        type: string
    type: object
...
  models.Farm:
    properties:
      address:
        description: Адрес
        type: string
      cowsCount:
        description: количество коров
        type: integer
      description:
        description: описание
        type: string
      districtId:
        description: ID района, в котором находится хозяйство
        type: integer
      email:
        description: эл. почта
        type: string
      hozNumber:
        description: |-
          Region   Region `json:"-"`
          RegionId uint
        type: string
      id:
        type: integer
      inn:
        description: ИНН
        type: string
      name:
        description: Название хозяйства
        type: string
      nameShort:
        description: Краткое название хозйства
        type: string
      parrentId:
        description: ID более управляющего хоз-ва (для хозяйства - холдинг, для фермы
          - хозяйство)
        type: integer
      phone:
        description: телефон
        type: string
      type:
        description: 'Тип: хозяйство, ферма, холдинг'
        type: integer
    type: object
...
  models.HoldRegisterRequest:
    properties:
      address:
        type: string
      cowsCount:
        type: integer
      description:
        type: string
      districtId:
        type: string
      email:
        type: string
      hozNumber:
        type: string
      id:
        type: integer
      inn:
        type: string
      name:
        type: string
      phone:
        type: string
      shortName:
        type: string
    type: object
  models.HozRegisterRequest:
    properties:
      address:
        type: string
      cowsCount:
        type: integer
      description:
        type: string
      districtId:
        type: integer
      email:
        type: string
      holdNumber:
        type: string
      hozNumber:
        type: string
      id:
        type: integer
      inn:
        type: string
      name:
        type: string
      phone:
        type: string
      shortName:
        type: string
    type: object
  models.Lactation:
    properties:
      abort:
        description: был ли аборт
        type: boolean
      calvingCount:
        description: 'Количество рожденных телят: 0 - мертворождение, 2 - двойня'
        type: integer
      calvingDate:
        $ref: '#/definitions/models.DateOnly'
      cowId:
        description: ID коровы, данные о лактации которой записаны
        type: integer
      days:
        description: количество дней, когда корова дает молоко
        type: integer
      fat305:
        description: Суммарный жир за 305 дней
        type: number
      fatAll:
        description: Суммарный жир
        type: number
      id:
        type: integer
      insemenationDate:
        allOf:
        - $ref: '#/definitions/models.DateOnly'
        description: Дата осеменения
      insemenationNum:
        description: Количество осеменений
        type: integer
      milk305:
        description: Суммарный надой за 305 дней
        type: number
      milkAll:
        description: Суммарный надой
        type: number
      number:
        description: номер лактации
        type: integer
      protein305:
        description: Суммарный белок за 305 дней
        type: number
      proteinAll:
        description: Суммарный белок
        type: number
      servicePeriod:
        description: 'сервис период коровы: время от отела до осеменения'
        type: integer
    type: object
...
  models.UserRegisterRequest:
    properties:
      email:
        type: string
      hozNumber:
        description: 'номер хоз-ва к которому привязвыается пользователь: либо существует,
          либо newHoz'
        type: string
      id:
        type: integer
      nameSurnamePatronimic:
        type: string
      password:
        type: string
      phone:
        type: string
      regionId:
        type: integer
      roleId:
        type: integer
    type: object
  user_create.createUserData:
    properties:
      newHold:
        allOf:
        - $ref: '#/definitions/models.HoldRegisterRequest'
        description: не обрабатывается
      newHoz:
        allOf:
        - $ref: '#/definitions/models.HozRegisterRequest'
        description: не обрабатывается
      newUser:
        allOf:
        - $ref: '#/definitions/models.UserRegisterRequest'
        description: данные пользователя для регистрации
    type: object
...
host: genmilk.ru
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: |-
    Сваггер сгенерирован из кода, поэтому может содержать неточности. По мере возможности они будут описаны далее
    Все даты передаются как строки
    Большая часть рутов не возвращает вложенные объекты
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: GenMilk API
  version: "1.0"
paths:
...
  /dailyMilks:
    get:
      description: Возвращает дойки удовлетворяющие фильтрам. Без фильтра не работает,
        т.к. слишком много доений
      parameters:
      - description: id lactation to search dailimilks
        in: query
        name: lactation_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.DailyMilk'
            type: array
        "404":
          description: Not Found
          schema:
            type: string
        "422":
          description: Unprocessable Entity
          schema:
            type: string
      summary: Get list of DailyMilks
      tags:
      - DailyMilks
  /dailyMilks/{id}:
    get:
      description: Возращает конкретную дойку коровы.
      parameters:
      - description: id of dailymilk
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DailyMilk'
        "404":
          description: Not Found
          schema:
            type: string
        "422":
          description: Unprocessable Entity
          schema:
            type: string
      summary: Get dailyMilk
      tags:
      - DailyMilks
...
  /farms:
    get:
      description: Возращает список ферм. Разрешает отсутсвие фильтров
      parameters:
      - description: ID более главной фермы, null для поиска холдингов
        in: query
        name: parrent_id
        type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Farm'
            type: array
        "404":
          description: Not Found
          schema:
            type: string
        "422":
          description: Unprocessable Entity
          schema:
            type: string
      summary: Get list of farms
      tags:
      - Farms
  /farms/{id}:
    get:
      description: Возращает конкретную ферму (хозяйство, холдинг)
      parameters:
      - description: id of farm to return
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Farm'
        "404":
          description: Not Found
          schema:
            type: string
        "422":
          description: Unprocessable Entity
          schema:
            type: string
      summary: Get farm
      tags:
      - Farms
  /lactations/{id}:
    get:
      description: Возращает конкретную лактацию
      parameters:
      - description: id лактации
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Lactation'
        "404":
          description: Not Found
          schema:
            type: string
        "422":
          description: Unprocessable Entity
          schema:
            type: string
      summary: Get list of farms
      tags:
      - Lactations
...
  /user/create:
    post:
      description: Рут для создания запроса на регистрацию
      parameters:
      - description: applied filters
        in: body
        name: userData
        required: true
        schema:
          $ref: '#/definitions/user_create.createUserData'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "422":
          description: Unprocessable Entity
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: User register request
      tags:
      - User
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"